[review] Remove BaseHTTPClient
