--- core/src/main/java/org/apache/iceberg/rest/BaseHTTPClient.java
@@ -1,211 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.iceberg.rest;
-
-import java.util.Map;
-import java.util.function.Consumer;
-import java.util.function.Supplier;
-import org.apache.iceberg.rest.HTTPRequest.HTTPMethod;
-import org.apache.iceberg.rest.auth.AuthSession;
-import org.apache.iceberg.rest.responses.ErrorResponse;
-
-/** A base class for {@link RESTClient} implementations using {@link HTTPRequest}. */
-public abstract class BaseHTTPClient implements RESTClient {
-
-  @Override
-  public abstract RESTClient withAuthSession(AuthSession session);
-
-  @Override
-  public void head(
-      String path, Supplier<Map<String, String>> headers, Consumer<ErrorResponse> errorHandler) {
-    HTTPRequest request = buildRequest(HTTPMethod.HEAD, path, null, headers.get(), null);
-    execute(request, null, errorHandler, h -> {});
-  }
-
-  @Override
-  public void head(String path, Map<String, String> headers, Consumer<ErrorResponse> errorHandler) {
-    HTTPRequest request = buildRequest(HTTPMethod.HEAD, path, null, headers, null);
-    execute(request, null, errorHandler, h -> {});
-  }
-
-  @Override
-  public <T extends RESTResponse> T delete(
-      String path,
-      Map<String, String> queryParams,
-      Class<T> responseType,
-      Supplier<Map<String, String>> headers,
-      Consumer<ErrorResponse> errorHandler) {
-    HTTPRequest request = buildRequest(HTTPMethod.DELETE, path, queryParams, headers.get(), null);
-    return execute(request, responseType, errorHandler, h -> {});
-  }
-
-  @Override
-  public <T extends RESTResponse> T delete(
-      String path,
-      Class<T> responseType,
-      Supplier<Map<String, String>> headers,
-      Consumer<ErrorResponse> errorHandler) {
-    HTTPRequest request = buildRequest(HTTPMethod.DELETE, path, null, headers.get(), null);
-    return execute(request, responseType, errorHandler, h -> {});
-  }
-
-  @Override
-  public <T extends RESTResponse> T delete(
-      String path,
-      Class<T> responseType,
-      Map<String, String> headers,
-      Consumer<ErrorResponse> errorHandler) {
-    HTTPRequest request = buildRequest(HTTPMethod.DELETE, path, null, headers, null);
-    return execute(request, responseType, errorHandler, h -> {});
-  }
-
-  @Override
-  public <T extends RESTResponse> T delete(
-      String path,
-      Map<String, String> queryParams,
-      Class<T> responseType,
-      Map<String, String> headers,
-      Consumer<ErrorResponse> errorHandler) {
-    HTTPRequest request = buildRequest(HTTPMethod.DELETE, path, queryParams, headers, null);
-    return execute(request, responseType, errorHandler, h -> {});
-  }
-
-  @Override
-  public <T extends RESTResponse> T get(
-      String path,
-      Class<T> responseType,
-      Supplier<Map<String, String>> headers,
-      Consumer<ErrorResponse> errorHandler) {
-    HTTPRequest request = buildRequest(HTTPMethod.GET, path, null, headers.get(), null);
-    return execute(request, responseType, errorHandler, h -> {});
-  }
-
-  @Override
-  public <T extends RESTResponse> T get(
-      String path,
-      Class<T> responseType,
-      Map<String, String> headers,
-      Consumer<ErrorResponse> errorHandler) {
-    HTTPRequest request = buildRequest(HTTPMethod.GET, path, null, headers, null);
-    return execute(request, responseType, errorHandler, h -> {});
-  }
-
-  @Override
-  public <T extends RESTResponse> T get(
-      String path,
-      Map<String, String> queryParams,
-      Class<T> responseType,
-      Supplier<Map<String, String>> headers,
-      Consumer<ErrorResponse> errorHandler) {
-    HTTPRequest request = buildRequest(HTTPMethod.GET, path, queryParams, headers.get(), null);
-    return execute(request, responseType, errorHandler, h -> {});
-  }
-
-  @Override
-  public <T extends RESTResponse> T get(
-      String path,
-      Map<String, String> queryParams,
-      Class<T> responseType,
-      Map<String, String> headers,
-      Consumer<ErrorResponse> errorHandler) {
-    HTTPRequest request = buildRequest(HTTPMethod.GET, path, queryParams, headers, null);
-    return execute(request, responseType, errorHandler, h -> {});
-  }
-
-  @Override
-  public <T extends RESTResponse> T post(
-      String path,
-      RESTRequest body,
-      Class<T> responseType,
-      Supplier<Map<String, String>> headers,
-      Consumer<ErrorResponse> errorHandler) {
-    HTTPRequest request = buildRequest(HTTPMethod.POST, path, null, headers.get(), body);
-    return execute(request, responseType, errorHandler, h -> {});
-  }
-
-  @Override
-  public <T extends RESTResponse> T post(
-      String path,
-      RESTRequest body,
-      Class<T> responseType,
-      Supplier<Map<String, String>> headers,
-      Consumer<ErrorResponse> errorHandler,
-      Consumer<Map<String, String>> responseHeaders) {
-    HTTPRequest request = buildRequest(HTTPMethod.POST, path, null, headers.get(), body);
-    return execute(request, responseType, errorHandler, responseHeaders);
-  }
-
-  @Override
-  public <T extends RESTResponse> T post(
-      String path,
-      RESTRequest body,
-      Class<T> responseType,
-      Map<String, String> headers,
-      Consumer<ErrorResponse> errorHandler,
-      Consumer<Map<String, String>> responseHeaders) {
-    HTTPRequest request = buildRequest(HTTPMethod.POST, path, null, headers, body);
-    return execute(request, responseType, errorHandler, responseHeaders);
-  }
-
-  @Override
-  public <T extends RESTResponse> T post(
-      String path,
-      RESTRequest body,
-      Class<T> responseType,
-      Map<String, String> headers,
-      Consumer<ErrorResponse> errorHandler) {
-    HTTPRequest request = buildRequest(HTTPMethod.POST, path, null, headers, body);
-    return execute(request, responseType, errorHandler, h -> {});
-  }
-
-  @Override
-  public <T extends RESTResponse> T postForm(
-      String path,
-      Map<String, String> formData,
-      Class<T> responseType,
-      Supplier<Map<String, String>> headers,
-      Consumer<ErrorResponse> errorHandler) {
-    HTTPRequest request = buildRequest(HTTPMethod.POST, path, null, headers.get(), formData);
-    return execute(request, responseType, errorHandler, h -> {});
-  }
-
-  @Override
-  public <T extends RESTResponse> T postForm(
-      String path,
-      Map<String, String> formData,
-      Class<T> responseType,
-      Map<String, String> headers,
-      Consumer<ErrorResponse> errorHandler) {
-    HTTPRequest request = buildRequest(HTTPMethod.POST, path, null, headers, formData);
-    return execute(request, responseType, errorHandler, h -> {});
-  }
-
-  protected abstract HTTPRequest buildRequest(
-      HTTPMethod method,
-      String path,
-      Map<String, String> queryParams,
-      Map<String, String> headers,
-      Object body);
-
-  protected abstract <T extends RESTResponse> T execute(
-      HTTPRequest request,
-      Class<T> responseType,
-      Consumer<ErrorResponse> errorHandler,
-      Consumer<Map<String, String>> responseHeaders);
-}
